Final | Banking Project | Account Service is Updated

+++ Final-BankingSystem/BankingSystem/src/main/java/com/example/bankingsystem/service/impl/CustomerServiceImpl.java
@@ -4,28 +4,40 @@ import com.example.bankingsystem.converter.CustomerConverter;
 import com.example.bankingsystem.dto.request.CustomerCreateRequestDTO;
 import com.example.bankingsystem.dto.request.CustomerUpdateRequestDTO;
 import com.example.bankingsystem.dto.response.CustomerGetResponseDTO;
+import com.example.bankingsystem.entity.Account;
 import com.example.bankingsystem.entity.Customer;
 import com.example.bankingsystem.exception.ServiceOperationAlreadyDeletedException;
+import com.example.bankingsystem.exception.ServiceOperationNotDeleteException;
 import com.example.bankingsystem.exception.ServiceOperationNotFoundException;
 import com.example.bankingsystem.repository.CustomerRepository;
+import com.example.bankingsystem.service.AccountService;
 import com.example.bankingsystem.service.CustomerService;
-import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Collection;
+import java.util.Objects;
 
 @Service
-@RequiredArgsConstructor
+
 public class CustomerServiceImpl implements CustomerService {
 
     private final CustomerRepository customerRepository;
     private final CustomerConverter customerConverter;
+    private final AccountService accountService;
+
+    public CustomerServiceImpl(CustomerRepository customerRepository, CustomerConverter customerConverter,@Lazy AccountService accountService) {
+        this.customerRepository = customerRepository;
+        this.customerConverter = customerConverter;
+        this.accountService = accountService;
+    }
 
     @Override
     public void addCustomer(CustomerCreateRequestDTO customerCreateRequestDTO) {
         Customer customer = customerConverter.toCustomer(customerCreateRequestDTO);
         customerRepository.save(customer);
-        //return new Result(true,"Customer is added successfully.");
+
 
     }
 
@@ -55,18 +67,22 @@ public class CustomerServiceImpl implements CustomerService {
     public String deleteCustomer(Long id, boolean hardDelete) {
         Customer customer = customerRepository.getById(id);
 
-        if (customer.isDeleted()) {
-            throw new ServiceOperationAlreadyDeletedException.CustomerAlreadyDeletedException("Customer is already deleted");
-        }
-        if (hardDelete) {
-            customerRepository.removeCustomerById(id);
+        Collection<Account> accountCollection = accountService.getAllAccountsofOneCustomer(id).stream().filter(i->(i.getBalance().compareTo(BigDecimal.ZERO) > 0)).toList();
+        if(Objects.isNull(accountCollection)) {
+
+            if (customer.isDeleted()) {
+                throw new ServiceOperationAlreadyDeletedException.CustomerAlreadyDeletedException("Customer is already deleted");
+            }
+            if (hardDelete) {
+                customerRepository.removeCustomerById(id);
+                return "Customer is deleted successfully";
+            }
+
+            customer.setDeleted(true);
+            customerRepository.save(customer);
             return "Customer is deleted successfully";
         }
-
-        customer.setDeleted(true);
-        customerRepository.save(customer);
-        return "Customer is deleted successfully";
-
+         throw new ServiceOperationNotDeleteException.CustomerBalanceNotZero("Customer has account that balance of it has money");
     }
 
     @Override
